import 8u171-b11

+++ modules/web/src/main/java/com/sun/javafx/webkit/UIClientImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, 2016, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2017, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -215,6 +215,14 @@ public final class UIClientImpl implements UIClient {
         return "";
     }
 
+    @Override public boolean canRunBeforeUnloadConfirmPanel() {
+        return false;
+    }
+
+    @Override public boolean runBeforeUnloadConfirmPanel(String message) {
+        return false;
+    }
+
     @Override public String[] chooseFile(String initialFileName, boolean multiple, String mimeFilters) {
         // get the toplevel window
         Window win = null;
@@ -312,7 +320,7 @@ public final class UIClientImpl implements UIClient {
     @Override public void startDrag(WCImage image,
         int imageOffsetX, int imageOffsetY,
         int eventPosX, int eventPosY,
-        String[] mimeTypes, Object[] values
+        String[] mimeTypes, Object[] values, boolean isImageSource
     ){
         content = new ClipboardContent();
         for (int i = 0; i < mimeTypes.length; ++i) if (values[i] != null) {
@@ -346,24 +354,26 @@ public final class UIClientImpl implements UIClient {
             //Image need to be created by target request only.
             //QuantumClipboard.putContent have to be rewritten in Glass manner
             //with postponed data requests (DelayedCallback data object).
-            Object platformImage = image.getWidth() > 0 && image.getHeight() > 0 ?
-                    image.getPlatformImage() : null;
-            if (platformImage != null) {
-                try {
-                    File temp = File.createTempFile("jfx", ".png");
-                    temp.deleteOnExit();
-                    ImageIO.write(
-                        toBufferedImage(Image.impl_fromPlatformImage(
-                            Toolkit.getToolkit().loadPlatformImage(
-                                platformImage
-                            )
-                        )),
-                        "png",
-                        temp);
-                    content.put(DataFormat.FILES, Arrays.asList(temp));
-                } catch (IOException | SecurityException e) {
-                    //That is ok. It was just an attempt.
-                    //e.printStackTrace();
+            if (isImageSource) {
+                Object platformImage = image.getWidth() > 0 && image.getHeight() > 0 ?
+                        image.getPlatformImage() : null;
+                if (platformImage != null) {
+                    try {
+                        File temp = File.createTempFile("jfx", ".png");
+                        temp.deleteOnExit();
+                        ImageIO.write(
+                            toBufferedImage(Image.impl_fromPlatformImage(
+                                Toolkit.getToolkit().loadPlatformImage(
+                                    platformImage
+                                )
+                            )),
+                            "png",
+                            temp);
+                        content.put(DataFormat.FILES, Arrays.asList(temp));
+                    } catch (IOException | SecurityException e) {
+                        //That is ok. It was just an attempt.
+                        //e.printStackTrace();
+                    }
                 }
             }
         }
